lab4/requirements.txt
psycopg2-binary==2.9.10
Jinja2==3.1.2
python-dotenv==1.0.0
